+++ pelican/tools/pelican_import.py
@@ -167,6 +167,35 @@ def dc2fields(file):
         yield (post_title, content, slugify(post_title), post_creadt, author, categories, tags, post_format)
 
 
+def posterous2fields(api_token, email, password):
+    """Imports posterous posts"""
+    import base64
+    import simplejson as json
+    import urllib2
+
+    def get_posterous_posts(api_token, email, password, page = 1):
+        base64string = base64.encodestring('%s:%s' % (email, password)).replace('\n', '')
+        url = "http://posterous.com/api/v2/users/me/sites/primary/posts?api_token=%s&page=%d" % (api_token, page)
+        request = urllib2.Request(url)
+        request.add_header("Authorization", "Basic %s" % base64string)
+        handle = urllib2.urlopen(request)
+        posts = json.loads(handle.read())
+        return posts
+
+    page = 1
+    posts = get_posterous_posts(api_token, email, password, page)
+    while len(posts) > 0:
+        posts = get_posterous_posts(api_token, email, password, page)
+        page += 1
+
+        for post in posts:
+            slug = post.get('slug')
+            if not slug:
+                slug = slugify(post.get('title'))
+            tags = [tag.get('name') for tag in post.get('tags')]
+            yield (post.get('title'), post.get('body_cleaned'), slug, post.get('display_date'), 
+                post.get('user').get('display_name'), [], tags, "html")
+
 def feed2fields(file):
     """Read a feed and yield pelican fields"""
     import feedparser
@@ -292,6 +321,8 @@ def main():
         help='Wordpress XML export')
     parser.add_argument('--dotclear', action='store_true', dest='dotclear',
         help='Dotclear export')
+    parser.add_argument('--posterous', action='store_true', dest='posterous',
+        help='Posterous export')
     parser.add_argument('--feed', action='store_true', dest='feed',
         help='Feed to parse')
     parser.add_argument('-o', '--output', dest='output', default='output',
@@ -308,6 +339,10 @@ def main():
         help='Disable storing slugs from imported posts within output. '
              'With this disabled, your Pelican URLs may not be consistent '
              'with your original posts.')
+    parser.add_argument('-e', '--email', dest='email',
+        help="Email address (posterous import only)")
+    parser.add_argument('-p', '--password', dest='password',
+        help="Password (posterous import only)")
 
     args = parser.parse_args()
 
@@ -316,10 +351,12 @@ def main():
         input_type = 'wordpress'
     elif args.dotclear:
         input_type = 'dotclear'
+    elif args.posterous:
+        input_type = 'posterous'
     elif args.feed:
         input_type = 'feed'
     else:
-        error = "You must provide either --wpfile, --dotclear or --feed options"
+        error = "You must provide either --wpfile, --dotclear, --posterous or --feed options"
         exit(error)
 
     if not os.path.exists(args.output):
@@ -333,6 +370,8 @@ def main():
         fields = wp2fields(args.input)
     elif input_type == 'dotclear':
         fields = dc2fields(args.input)
+    elif input_type == 'posterous':
+        fields = posterous2fields(args.input, args.email, args.password)
     elif input_type == 'feed':
         fields = feed2fields(args.input)